LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-next-focus-expected.txt

Before
Section 1
Section 2  Focusable
Section 3
After

FAIL CSS Test: next focus search start after ::scroll-marker activation is from its ultimate originating element assert_equals: the first ::scroll-marker gets activated upon clicking expected "rgb(0, 128, 0)" but got ""
